Move Defaultable into GraphQL.API
Also removes unused DefaultArgument.

Fixes #149. Fixes #148.
